Update docstrings for doctest

+++ gymnasium/spaces/tuple.py
@@ -19,7 +19,7 @@ class Tuple(Space[typing.Tuple[Any, ...]], typing.Sequence[Any]):
         >>> from gymnasium.spaces import Tuple, Box, Discrete
         >>> observation_space = Tuple((Discrete(2), Box(-1, 1, shape=(2,))), seed=42)
         >>> observation_space.sample()
-        (0, array([-0.3991573 ,  0.21649833], dtype=float32))
+        (np.int64(0), array([-0.3991573 ,  0.21649833], dtype=float32))
     """
 
     def __init__(
